==> 05_DEG_Outline.Rmd <==
---
title: "BOO 2025 - Example Analysis"
subtitle: "Script 5: Differential expression - Questions"
date: "`r Sys.Date()`" 
author: 
 Pien van Zon
output:
  html_document:
    code_download: true
    theme: united
    lowlight: tango
    df_print: paged
    toc: true
    number_sections: true
    toc_float:
      collapsed: false
      smooth_scroll: true
---

***

> In this script, you will identify differentially expressed genes (DEGs) and visualize your results.

***

# Setup

## Clean

As before, we perform several setup steps at the start of the script to ensure our work is reproducible and clear. 

**Exercise 1: Download the R markdown template, clean your environment, and set the following variables:**

* `root_dir` - project folder
* `count_path` - location of the `countData` object
* `metadata_path` - location of the `metaData` object
* `low_store` - location to store DEG results for the low toxicity compound
* `high_store` - location to store DEG results for the high toxicity compound

```{r clean}
# This clears the environment
rm(list=ls())

root_dir <- "C:/Users/pienz/Universiteit Leiden/BOO 2025 - BOO CDS Giulia team - BOO CDS Giulia team/Students/Pien/project/pienvanzon_BOO//"  

count_path <- "countData.Rdata"
metadata_path <- "metaData.Rdata" 

low_store <- "C:/Users/pienz/Universiteit Leiden/BOO 2025 - BOO CDS Giulia team - BOO CDS Giulia team/Students/Pien/project/pienvanzon_BOO/output/"
high_store <- "C:/Users/pienz/Universiteit Leiden/BOO 2025 - BOO CDS Giulia team - BOO CDS Giulia team/Students/Pien/project/pienvanzon_BOO/output/"

```

***

## Packages

In this script, we introduce a new package: `DESeq2`, a package that analyses count-based sequencing data, such as TempO-Seq, to identify differentially expressed genes between experimental conditions.

**Exercise 2: Load `DESeq2`, `ggrepel`, and `tidyverse` into your R environment:**

```{r load-packages, warning=F, message=F}
# This loads the packages
library(tidyverse)
library(ggrepel)
library(DESeq2)

```

***

## Load data

**Exercise 3: Load the count and metadata into your R environment:**

```{r load-data, warning=FALSE, message=FALSE}
#
load(count_path)
load(metadata_path)
```

***

# DEG Analysis 

## Mean ID

In order to perform DEG analysis we need to ensure we have the correct `mean_ID` and `mean_ID_control` variables. These can be used to tell DESeq2 which samples to compare:

* `mean_ID` is a group identifier, describing which treatment condition each sample belongs to
* `mean_ID_control` is a control identifier, describing which control samples a group should be compared to

In the case of our data we will have:

* A `mean_ID` level for each experimental condition (6 concentrations for 2 compounds plus two plate-specific DMSO groups)
* No `mean_ID_control` for the DMSO controls, but the appropriate DMSO control specified for the treated samples

**Exercise 4: Identify these variables in your `metaData`:**

```{r inspect}
#
metaData %>% 
  select(mean_ID, mean_ID_control) %>% 
  filter(!is.na(mean_ID_control)) %>% 
  distinct()
```

***

## Contrast

DESeq2 uses an object called a contrast, which details the comparisons that will be run. We often want to save several comparisons together in one object, such as all results for each compound together.

**Exercise 5: Create a `contrast` data frame, including a variable `file_ID` which describes which file each comparison should be saved in:**

```{r contrast}
#
contrast <- metaData %>% 
  filter(!is.na(mean_ID_control)) %>% 
  select(mean_ID, mean_ID_control, file_ID = compound_ID) %>% 
  distinct

contrast

```

***

## Subsets

We want to perform DEG analysis for each concentration of the low and high toxicity compounds.

**Exercise 6: Subset `contrast` to include only the comparisons for the low toxicity compound and create a vector `levels` which can be used to loop through the selected `mean_ID`:**


```{r levels}
#
contrast_low <- contrast %>% 
  filter(file_ID == "CS1131")

contrast_low

```

```{r}
levels <- unique(contrast_low$mean_ID)

levels
```
***

Before looping through `levels`, let's go through the steps required to perform a DEG analysis.

**Exercise 7: Using the first value in `levels`, subset the `contrast`, `countData`, and `metaData` to include only the relevant samples:**

<details>
  <summary><strong>Hint</strong></summary>

  Make sure you create new objects, such as `countData_sub` and `metaData_sub` to prevent overwriting the full data frame.

</details>


```{r subset}
# selecting the first mean ID
i <- levels[1]
i

```
```{r}
# Subset the contrast
contrast_sub <- contrast_low %>% 
  filter(mean_ID == i)
contrast_sub
```

```{r}
# Subset the metadata
metaData_sub <- metaData %>% 
  filter(mean_ID %in% c(contrast_sub$mean_ID, contrast_sub$mean_ID_control))
metaData_sub
```

```{r}
# subset the count data
countData_sub <- countData[ , metaData_sub$sample_ID]
countData_sub[1:10, ]
```



***

## Fit models

Now that we have subset our data, we can use it to create a DESeq object (`dds`). This also requires us to set our `design`, or model formula. In our data, this will be a simple comparison of two levels of `mean_ID`.

**Exercise 8: Use the `DESeqDataSetFromMatrix()` function to create a `dds` object:**

<details>
  <summary><strong>Hint</strong></summary>

  Designs are a special type of object in R, the formula, which is specified using a tilde (e.g. `~ mean_ID`).

</details>


```{r fit}
#
dds <- DESeqDataSetFromMatrix(
  countData = countData_sub,
  colData = metaData_sub,
  design = ~ mean_ID
)

```

***

Next, we use the `DESeq()` function to:

1. Calculate **size factors** using the median ratio method, which normalizes for differences in sequencing depth or library size across samples. The motivation is similar to CPM normalization, but each sample is instead compared to a pseudo-reference sample composed of the geometric mean of all genes. 
2. Estimate **dispersions** (biological variability) for each gene across replicates, in order to properly quantify confidence in results (which will be lower for highly variable genes).
3. Fit a **generalized linear model** (GLM) for each gene, using a negative Binomial distribution. This takes the normalized counts (from size factors), design formula, and estimated dispersion to produce log2 fold changes (log2FC), standard errors (SEs), and p-values.

The output object, `dds`, contains all the model fits and is ready for extracting results.

**Exercise 9: Use the `DESeq()` function to update the `dds` object and include all specified model fits:**

```{r deseq}
#
dds <- DESeq(dds)

```

```{r}
dds
```


***

## Extract results

Results can be extracted from a `dds` object using the `results()` function. However, we additionally apply shrinkage to our results. **Shrinkage** pulls extreme or noisy log2FC towards zero when they are not well supported by the data, usually due to:

* Low expression levels
* High dispersion (biological variability)
* Small sample sizes

Shrunk log2FC are more stable and realistic, especially in these scenarios. **Normal** shrinkage uses a Normal prior distribution (bell curve with its peak at 0), which assumes that most genes are not differentially expressed and encourages smaller fold changes unless there is strong evidence to the contrary. 

The final log2FCs are then a posterior estimate that combines:

* The observed data, and
* The prior (Normal distribution centered at 0)

**Exercise 10: Use `lfcShrink()` to extract results for the selected comparison:**

<details>
  <summary><strong>Hint</strong></summary>

  You will get a warning using normal shrinkage, but our experience suggests that it is the best method for these simpler designs.

</details>


```{r shrinkage}
#
res_low <- lfcShrink(dds, 
                 contrast = c("mean_ID",
                              contrast_sub$mean_ID,
                              contrast_sub$mean_ID_control),
                 type="normal")

```
This reduces noise from the low expressed genes.

```{r}
head(res_low)
```
The differential expression comparison between CS1131_1_P6 and DMSOLow_P6 shows no significant DEGs among the top genes. Fold changes are small, and adjusted p-values (padj) are high or unavailable. This suggests low biological effect or high within-group variability at this specific concentration for this compound.

***

**Exercise 11: Wrangle the output to additional columns:**

* `gene_symbol` containing the gene names
* `mean_ID` containing the mean ID of the treated samples
* `mean_ID_control` containing the mean ID of the control samples

```{r results}
#
res_low <- data.frame(res_low) %>% 
  rownames_to_column(var = "gene_symbol") %>% 
  mutate(
    mean_ID = contrast_sub$mean_ID,
    mean_ID_control = contrast_sub$mean_ID_control
  )

head(res_low)


```

***

## Loop

Using a `for` loop, you can run each of these steps for every required comparison.

**Exercise 12: Loop through the remaining values of the `levels` variable to add all of the comparisons for the low toxicity compound to your `res` object:**

```{r loop}
#
for(i in levels[-1]){
  contrast_sub <- contrast_low %>% 
    filter(mean_ID == i)
  
  metaData_sub <- metaData %>% 
    filter(mean_ID %in% c(contrast_sub$mean_ID, contrast_sub$mean_ID_control))
  
  countData_sub <- countData[ , metaData_sub$sample_ID]

  dds <- DESeqDataSetFromMatrix(
    countData = countData_sub,
    colData = metaData_sub,
    design = ~ mean_ID)  
  
  dds <- DESeq(dds)
  
  res_low <- rbind(res_low, 
    data.frame(lfcShrink(dds, 
                 contrast = c("mean_ID",
                              contrast_sub$mean_ID,
                              contrast_sub$mean_ID_control),
                 type="normal")) %>% 
    rownames_to_column(var = "gene_symbol") %>% 
    mutate(
      mean_ID = contrast_sub$mean_ID,
      mean_ID_control = contrast_sub$mean_ID_control))
}

```

```{r}
head(res_low)
```

```{r}
dim(res_low)
```


***

**Exercise 13: Perform DEG analysis for the high toxicity compound, saving the output in `res_high`:**

First, we make the contrast for the high toxicity compound and save the levels:

```{r high-contrast}
#
contrast_high <- contrast %>% 
  filter(file_ID == "CS1119")

contrast_high
```

```{r}
levels <- unique(contrast_high$mean_ID)

levels
```


Then, we loop through `levels` and save results:

```{r high-loop}
#
res_high <- data.frame()

for(i in levels){
  contrast_sub <- contrast_high %>% 
    filter(mean_ID == i)
  
  metaData_sub <- metaData %>% 
    filter(mean_ID %in% c(contrast_sub$mean_ID, contrast_sub$mean_ID_control))
  
  countData_sub <- countData[ , metaData_sub$sample_ID]

  dds <- DESeqDataSetFromMatrix(
    countData = countData_sub,
    colData = metaData_sub,
    design = ~ mean_ID)  
  
  dds <- DESeq(dds)
  
  res_high <- rbind(res_high, 
    data.frame(lfcShrink(dds, 
                 contrast = c("mean_ID",
                              contrast_sub$mean_ID,
                              contrast_sub$mean_ID_control),
                 type="normal")) %>% 
    rownames_to_column(var = "gene_symbol") %>% 
    mutate(
      mean_ID = contrast_sub$mean_ID,
      mean_ID_control = contrast_sub$mean_ID_control))
}

```

```{r}
head(res_high)
```

```{r}
dim(res_high)
```


***

# Defining DEGs

## Thresholds

There are many ways to define a differentially expressed gene based either on the log2 fold change (log2FC, which reflects the effect size), the adjusted p-value (which reflects the statistical significance after correcting for multiple testing), or a combination of both. For example, some studies may define a gene as differentially expressed if it has an adjusted p-value below 0.05, indicating high statistical confidence, while others may require an absolute log2FC above log2(1.2) (i.e., a twenty percent change) to ensure the change is also biologically meaningful. Combining both criteria helps balance sensitivity and specificity, reducing the chance of identifying genes that are statistically but not biologically relevant, or vice versa.

**Exercise 14: Create two new variables `dir` and `sig` that describe each gene's significance (T/F) and direction of effect (up/down):**

<details>
  <summary><strong>Hint</strong></summary>

  Don't forget to do it for the high and low toxicity compounds.

</details>

```{r sig-dir}
#
res_high <- res_high %>% 
  mutate(
    dir = ifelse(log2FoldChange < 0, "Down", "Up"),
    sig = ifelse(padj <= 0.05 & abs(log2FoldChange) > log2(1.2), TRUE, FALSE)
  )

table(res_high$sig, res_high$dir, useNA="always")

```
Based on an adjusted p-value threshold of 0.05 and a minimum absolute log2 fold change of log2(1.2), we identified 4,180 significantly differentially expressed genes, of which 2,046 were downregulated and 2,134 were upregulated. The majority of genes showed no significant change, and ~5,354 genes are excluded due to insufficient data for statistical testing for the high tox compound.

```{r}
res_low <- res_low %>% 
  mutate(
    dir = ifelse(log2FoldChange < 0, "Down", "Up"),
    sig = ifelse(padj <= 0.05 & abs(log2FoldChange) > log2(1.2), TRUE, FALSE)
  )

table(res_low$sig, res_low$dir, useNA="always")
```
For the low toxicity compound, we identified 4,070 significantly differentially expressed genes, including 2,140 downregulated and 1,930 upregulated. The overall DEG count is comparable to the high toxicity compound, though the direction of change differs slightly. Most genes were not significantly changed, and a small subset (~2,100) was excluded due to low expression or test reliability.

***

## Save

It is important to save output of analyses locally, especially when they can be quite lengthy to rerun. This will save you considerable time later in the project. 

**Exercise 15: Save the results using the path variables you created at the start of the script:**

```{r save}
# This saves the results
save(res_low, file=file.path(root_dir, "output/res_low_05_DEG.RData"))
save(res_high, file=file.path(root_dir, "output/res_high_05_DEG.RData"))

```

***

# Plots

## Dose-response plots

One of the plots that is useful to make following DEG analysis is visualization of any dose response effects. In order to ensure that all data point are plotted correctly, it is helpful to create a base data frame for all the direction, concentration combinations.

**Exercise 16: Create a data frame `plot_df` with all compound, concentration, and direction combinations:**

<details>
  <summary><strong>Hint</strong></summary>

  It may be helpful to `rbind()` your high and low toxicity compound results together for visualizations.

</details>

```{r ref-df}
#
res <- rbind(res_high, res_low)

plot_df <- res %>% 
  filter(!is.na(dir)) %>% 
  select(mean_ID, dir) %>% 
  distinct()

plot_df
```

***

Using `mean_ID` and `separate()`, you can extract information about the compound and concentration of each experimental group.

**Exercise 17: Create a data frame containing the number of DEGs, concentration, and compound, for each experimental condition and direction:**

<details>
  <summary><strong>Hint</strong></summary>

  Don't forget to set the number of DEGs to 0, where appropriate. 

</details>

```{r plot-df}
#
plot_df <- res %>% 
  filter(sig == TRUE & !is.na(dir)) %>% 
  group_by(mean_ID, dir) %>%
  summarize(n = n(), .groups = "drop") %>% 
  right_join(y = plot_df, by = c("mean_ID", "dir")) %>% 
  mutate(n = ifelse(is.na(n), 0, n)) %>% 
  separate(mean_ID, sep="_", into=c("Compound", "Conc_ID", NA), remove=F)

plot_df
```

***

**Exercise 18: Plot the number of DEGs across concentrations for the two compounds. Do you see a dose-response?**

```{r dose-response}
#
plot_df %>% 
  mutate(n = ifelse(dir == "Down", -n, n)) %>% 
  ggplot(aes(x=Conc_ID, y=n, fill=dir)) +
  geom_col() +
  geom_hline(aes(yintercept=0)) +
  theme_bw() +
  facet_wrap(~Compound, scales="free_x")

```
This plot shows a distinct difference in response profiles between the two compounds. CS1119 (high toxicity) induces a strong transcriptional response at all concentrations, while CS1131 (low toxicity) shows a progressive increase in DEGs with dose — consistent with a dose-dependent response. This supports the expected toxicological profiles and helps validate your experimental design.

***

## Volcano plots

A volcano plot is useful to visualize effect sizes and significance of each gene in one plot, by plotting log2 fold changes (x-axis) against -log10 p-values (y-axis).

**Exercise 19: Create volcano plots for each of the compounds:**

<details>
  <summary><strong>Hint</strong></summary>

  If you want to plot all concentrations, it may be sensible to create distinct plots for the two compounds. However, you could also plot only the top concentration volcano plot, if the dose response is suggestive that this would be more informative. 

</details>

```{r volcano}
#
for(i in unique(metaData$compound_ID)[1:2]) {
  print(res %>% 
          separate(mean_ID, sep="_", into=c("Compound", "Conc_ID", NA), remove=F) %>% 
          filter(!is.na(sig) & Compound == i) %>% 
          mutate(log10p = -log10(pvalue)) %>% 
          ggplot(aes(x = log2FoldChange, y = log10p)) +
          geom_point(aes(color=sig), size=0.5) +
          geom_text_repel(aes(label = ifelse(padj < 1E-4, gene_symbol, NA)),
                    color="dodgerblue", size=2) +
          xlim(-max(abs(res$log2FoldChange), na.rm=T), max(abs(res$log2FoldChange), na.rm=T)) +
          scale_color_manual(values = c("grey", "dodgerblue")) +
          theme_bw() + 
          ggtitle(i) +
          facet_wrap(~Conc_ID))
}

```

Volcano plots of CS1119 across seven concentrations show widespread differential gene expression. Strong transcriptional effects are observed at both low (Conc_ID 1–2) and high doses (Conc_ID 6–7), with many significant DEGs highlighted. This supports a non-linear but robust transcriptional response, characteristic of a high-toxicity compound.

Volcano plots for CS1131 show minimal differential expression at low and mid concentrations (Conc_ID 1–6), with a pronounced increase in both up- and downregulated genes at the highest dose (Conc_ID 7). This supports a dose-dependent transcriptional response, consistent with CS1131’s lower toxicity profile compared to CS1119.


***

## MA plots

One feature that volcano plots miss is the baseline expression of genes. This can mean that very lowly expressed genes appear important (due to high significance and effect sizes). Visualizing results using an MA plot, which plots log2 base mean (y-axis) against log2 fold changes (x-axis), can be helpful to distinguish between noisy genes and biologically relevant ones.

**Exercise 20: Create MA plots for the two compounds. Outline your findings:**

```{r ma}
#
for(i in unique(metaData$compound_ID)[1:2]) {
  print(res %>% 
          separate(mean_ID, sep="_", into=c("Compound", "Conc_ID", NA), remove=F) %>% 
          filter(!is.na(sig) & Compound == i) %>% 
          mutate(log2baseMean = log2(baseMean)) %>% 
          ggplot(aes(x = log2baseMean, y = log2FoldChange)) +
          geom_point(aes(color=sig), size=0.5) +
          geom_text_repel(aes(label = ifelse(padj < 1E-4, gene_symbol, NA)),
                    color="dodgerblue", size=2) +
          ylim(-max(abs(res$log2FoldChange), na.rm=T), max(abs(res$log2FoldChange), na.rm=T)) +
          scale_color_manual(values = c("grey", "dodgerblue")) +
          theme_bw() + 
          ggtitle(i) +
          facet_wrap(~Conc_ID))
}

```
The MA plots for CS1119 show a strong and consistent transcriptional response across most concentrations, especially at 1, 4, and 6. Significant genes are primarily found at moderate to high expression levels, suggesting a robust and biologically meaningful response rather than noise from lowly expressed genes. This further supports CS1119’s high toxicity profile.

CS1131 shows minimal differential expression at low and intermediate concentrations, with DEGs mainly appearing in moderately expressed genes. However, at the highest concentration (Conc_ID 7), the response becomes more pronounced and reliable, with widespread changes involving well-expressed genes. This supports a dose-dependent response, typical for non-toxic or mildly active compounds.

***

# Advanced question

## Gene level plots

It can also be useful to plot individual gene results, especially to investigate gene-level dose responses.

**Exercise 21: Plot the top 16 genes for each of the compounds in a dose-response plot. Are there any genes with particular relevance to hepatotoxicity?**

```{r high-gene}
#

```

***

# Session Info

```{r session-info}
#

```

***

This is the end of the differential gene expression analysis. Next, please move on to functionally interpret your results using over-representation analysis in `06_ORA_Outline.Rmd`. Example answers will be available from the GitHub at EOB on Friday.

***
